Add files via upload
Updated player movement script to have pull method and made small changes that I marked with my name next to it to make the pull work and fix the turning when jumping issue. Also the pulling box script that will need to be added to any item you want to push or pull.