Burn ability + Torches!
Added the ability to manually burn objects at the press of the F key. Added a prefab for torches that light when this burn function is used nearby. These torches have an optional "Glide area" added to them for the torch glide puzzle. Wick requires to have a trigger area added to him as a child which acts as the burn area.