Add files via upload
Made changes to the flipsprite, respond, update, pull, and stop pulling functions to fix the issue with the respawning in the air if holding a crate during respawn. Also added a crate reset script so when the player respawns then so do the crates getting rid of the issue where if you fall off with the crate the crate disappears.